+++ helper/nr-helper.cc
@@ -388,8 +388,27 @@ NrHelper::CreateUePhy(const Ptr<Node>& n,
     channelPhy->SetIsGnb(false);
     channelPhy->SetDevice(dev); // each NrSpectrumPhy should have a pointer to device
 
-    auto antenna = m_ueAntennaFactory.Create(); // Create antenna object
-    channelPhy->SetAntenna(antenna);
+    bool usingUniformPlanarArray =
+        m_ueAntennaFactory.GetTypeId() == UniformPlanarArray::GetTypeId();
+    // Create n antenna panels and beam manager for Ue
+    for (auto i = 0; i < channelPhy->GetNumPanels(); i++)
+    {
+        auto antenna = m_ueAntennaFactory.Create(); // Create antenna object per panel
+        channelPhy->AddPanel(antenna);
+        // Check if the antenna is a uniform planar array type
+        if (usingUniformPlanarArray)
+        {
+            Ptr<BeamManager> beamManager = m_ueBeamManagerFactory.Create<BeamManager>();
+            auto uniformPlanarArray = DynamicCast<UniformPlanarArray>(antenna);
+            beamManager->Configure(uniformPlanarArray);
+            channelPhy->AddBeamManager(beamManager);
+        }
+    }
+    if (usingUniformPlanarArray)
+    {
+        // Config bearing angles for all panels installed in NrSpectrumPhy
+        channelPhy->ConfigPanelsBearingAngles();
+    }
 
     cam->SetNrSpectrumPhy(channelPhy); // connect CAM
 
@@ -445,14 +464,6 @@ NrHelper::CreateUePhy(const Ptr<Node>& n,
     channelPhy->SetPhyRxDataEndOkCallback(MakeCallback(&NrUePhy::PhyDataPacketReceived, phy));
     channelPhy->SetPhyRxCtrlEndOkCallback(phyRxCtrlCallback);
     channelPhy->SetPhyRxPssCallback(MakeCallback(&NrUePhy::ReceivePss, phy));
-    // Check if the antenna is a uniform planar array type
-    auto uniformPlanarArray = DynamicCast<UniformPlanarArray>(antenna);
-    if (uniformPlanarArray)
-    {
-        Ptr<BeamManager> beamManager = m_ueBeamManagerFactory.Create<BeamManager>();
-        beamManager->Configure(uniformPlanarArray);
-        channelPhy->SetBeamManager(beamManager);
-    }
     phy->InstallSpectrumPhy(channelPhy);
     return phy;
 }